requirements.txt
openai
PyPDF2
pandas
python-multipart
streamlit
python-docx